some refactor and clean up

+++ app/src/main/java/com/example/bookfinder/model/Book.kt
@@ -1,6 +1,5 @@
 package com.example.bookfinder.model
 
-import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -26,7 +25,7 @@ data class Volume(
     val publishedDate: String? = null,
     val description: String? = null,
     val pageCount: Int? = null,
-    val averageRating: Int? = null,
+    val averageRating: Double? = null,
     val ratingsCount: Int? = null,
     val imageLinks: Map<String, String>? = null,
     val language: String? = null,
